requirements.txt
Package            Version
------------------ -----------
bcrypt             4.3.0
blinker            1.9.0
certifi            2025.1.31
charset-normalizer 3.4.1
click              8.1.8
colorama           0.4.6
Flask              3.1.0
Flask-Bcrypt       1.0.1
Flask-MySQL        1.6.0
idna               3.10
itsdangerous       2.2.0
Jinja2             3.1.6
MarkupSafe         3.0.2
pip                25.0.1
PyMySQL            1.1.1
python-dateutil    2.9.0.post0
python-dotenv      1.1.0
requests           2.32.3
six                1.17.0
urllib3            2.3.0
Werkzeug           3.1.3
